refactor: separated login element and sign up element of profile page
