+++ frontend/src/types/tournament.ts
@@ -14,7 +14,7 @@ export interface HostProfile {
 }
 
 export interface Tournament {
-  id?: number;
+  id: number;
   name: string;
   startDateTime: string;
   endDateTime: string;
